Clean up mixfile metadata
In preparation for publishing to Hex, there were a few things to clean
up:

- removed the unused, auto-generated config/ stuff
- add ex_doc dev dependency (I presume I need this so that the hex.pm
  page will link to the hexdocs.pm page properly?)
- add package metadata: files, license, links, etc
- tszuj up the API docs: link source to github

The API docs aren't very useful, but including the README.md in the
ex_doc output doesn't work very well because of my uses of GitHub emoji.
Hoping people treat the hex.pm pages the same way I do: find it so that
I can immediately follow the link to GitHub for the "real" docs/source.