Fix application dependencies
Per further discussion on #1, a deployed Erlang app needs to know of all
the library applications, even if they aren't active.
RemoteIp.Headers.Forwarded was breaking at runtime from such apps
because Combine was not being loaded (even though it doesn't have a
supervisor to start or anything).

I believe this would be fixed by adding everything to `:applications`,
which Elixir 1.4 would do anyway, but for the sake of specificity I've
added all of remote_ip's dependencies (all the other 2 of them...) to
their "proper" places.