expose the last_forwarded_ip/2 function for usage outside of Plug
In some circumstances, it is possible to get a subset of headers without the
full conn. An example of this is in Phoenix sockets, where only x_headers can be
fetched.

https://hexdocs.pm/phoenix/Phoenix.Endpoint.html#socket/3-shared-configuration

In this case, currently using RemoteIP involves doing something like this as
conn is not available:

```elixir
def connect({connect_info: %{x_headers: headers}}) do
  RemoteIp.call(%Plug.Conn{req_headers: headers}, RemoteIp.init()).remote_ip
end
```

With this PR, someone using this library with Phoenix sockets could do:

```elixir
  RemoteIp.last_forwarded_ip(headers)
```