Add failing Logger.metadata tests
The problem with the implementation from #22 is that it fails to set the
logger metadata when an IP can't be parsed out of the forwarding
headers. I think it's safe to say that the less surprising behavior
would be to go ahead and add the metadata anyway, even though it's
coming from the original :remote_ip and not the :req_headers.

These tests have been deliberately added to enforce this expected
behavior, although the implementation currently fails.