Stop converting :headers to a MapSet
I've long suspected that the use of MapSet was a solution in search of a
problem. I guess I was just trying to be cool or something when I
started using it. 🤷

On paper, the membership check is O(1) vs O(n) in the size of the
:headers option. But I worried that, in practice, n would be so small
that it might actually *hurt* to use MapSet. I could easily imagine the
dominant term in a membership check being the hash operations (or
whatever Erlang's :maps uses) rather than the string comparisons.

Realistically, I figure you're configuring :headers using either

* the defaults, of which there are only 4 (forwarded, x-forwarded-for,
  x-client-ip, x-real-ip); or

* your own bespoke header, of which there will probably only be one,
  right? I can't imagine going off page by an ungodly number of headers,
  at any rate.

And it's probably also safe to assume you're not going to load up the
configuration with so many duplicates that the MapSet data structure
starts making sense. (If you do, you kind of reap what you sow.)

Furthermore, long after I'd gone with using MapSet, the function
RemoteIp.from/2 was added in e42a8ccd992a1eda62935f9b1e481fd31d76bb8c.
Before, RemoteIp.init/1 was generally called only at compile time
(unless you set the Plug.Builder :init_mode option, but that's
uncommon). But with the new function, we have to convert :headers into a
MapSet at runtime for *each* call! That's probably a much more
meaningful overhead than the membership checks.

So in the back of my mind I always wondered what the real difference
was. Now I've finally benchmarked it:
https://gist.github.com/ajvondrak/b0445fb9c70ef135cc224b22a674ec41

The results seem to confirm all of my suspicions, at least when using
the default :headers configuration.

* Universally, the runtime conversion from List to MapSet suffers a lot.

* The only case where MapSet seemed quicker than Enum (on this run) was
  when getting a hit on the first element, "forwarded". But the
  difference was so small as to effectively be a rounding error.

* All the other tests saw MapSet being slower than Enum by a noticeable
  margin. On an example of headers taken from one of normal my browser
  requests, using MapSet was 28% slower than Enum!

It seems the coffin is full of nails. I'm sure this micro-optimization
is still a wash in practice, but it also just makes the code simpler,
and that's always a win.