Add .tool-versions file
This is the version management file used by [asdf](https://asdf-vm.com).
It should help developers on this repo use the same version of
Elixir/OTP locally, whether directly (via asdf) or indirectly (by
telling them what they ought to be using). It's a similar idea to how
mix.lock is checked into the repo: it's for the benefit of local devs,
not for the version constraints of the remote_ip library itself.