Begin refactoring raw debug messages into id-based system
Rather than openly coding strings into all the RemoteIp.Debug.log/2 call
sites (especially as formatting gets increasingly complex), I'd prefer
to abstract over the type of message we're going to be logging. Then we
won't be stuck in quite as rigid a "#{msg}: #{output}" format, and the
call sites can be made even simpler.

In their current state, some of the message prefixes are actually
dynamically interpolated strings that range over the inputs to the
expression in the do-block. For example,

  RemoteIp.Debug.log("#{inspect(ip)} is a reserved IP") do
    @reserved |> contains?(ip)
  end

generates a message like

  {127, 0, 0, 1} is a reserved IP: true

So, the idea here is to generalize the old RemoteIp.Debug.log/2 macro
(that takes in a raw string prefix) into the new RemoteIp.Debug.log/3
macro that takes in an abstract message id, optional input values, and
the do-block to compute the output. The actual log messages can then be
generated by logic in the RemoteIp.Debug module.

We start doing this by moving the Logger.debug/2 call into the
RemoteIp.Debug.__log__/3 function, where we'll be able to compute a log
message at runtime. As of yet, we only have a message_for/3 clause that
maintains backwards-compatibility with the raw strings currently being
passed into the log/2 call sites. Since the inputs are optional, all the
arity-2 call sites should actually still work. So all the code and tests
maintain their old behavior, and the docs don't need to be updated yet.
But we'll be able to iteratively refactor each call site onto the new
system one step at a time.

Since we moved the actual Logger.debug/2 call site from the macro
expansion into the runtime RemoteIp.Debug.__log__/3 call, we don't need
the __using__/1 callback anymore. We can just require the Logger macros
from within the RemoteIp.Debug module itself.