Use new-style RemoteIp.Debug logging for plug options
We can now see an advantage to using this id-based system: it's much
easier to perform arbitrary string manipulation whose logic is housed
entirely in the RemoteIp.Debug module. Here, we make the CIDR notation
readable again. This undoes the work of InetCidr.parse/1, true... But
after all, this logging is for debugging purposes. Human readability
comes at the cost of some computation that you probably shouldn't be
enabling in prod anyway.