Use new-style RemoteIp.Debug logging for RemoteIp.{call,from}/2
The message IDs are separated so that we can provide useful information
about the "before" state in the RemoteIp.call/2 case. This is made
easier by having all the formatting logic in the RemoteIp.Debug module,
versus trying to cram it into some interpolated string prefix with the
old-style logging.