Remove last use of old-style RemoteIp.Debug logging
This message is rendered redundant by the :from & :call messages. It's
also the only other use of raw strings, so we can remove that clause
from RemoteIp.Debug.message_for/3.

This officially makes the log/3 docs outdated. Since we have `@moduledoc
false` anyway, I'm just going to remove the `@doc` for now. Eventually,
we should write public-facing module docs that explain how to
enable/disable the logging, but there's still work to be done hammering
out the implementation of RemoteIp.Debug.

The default clause of message_for/3 also remains intact, which keeps the
test suite happy. But it should also be removed in the future when I get
around to writing more specific tests.