Add failing tests for compile-time log purging
This test would've been fine if we only used Logger.debug/2, since the
level there is always known. But now the RemoteIp.Debug.__log__/3
function cannot pass in a literal atom to Logger.log/3 because we have
to look it up in the config first with Application.get_env/3. Even with
a module attribute, the Kernel.@/1 macro would need to be expanded
before the Logger.log/3 call.

Therefore, having configurable levels has broken level-based
compile-time log purging. While not a direct regression against older
docs of remote_ip that instructed people to purge [application:
:remote_ip], the behavior is still potentially surprising. Besides, I
want to avoid breaking any configs that were inspired by those docs (to
the extent that that's practical).