Teach RemoteIp.Debug to customize which messages get logged
The main advantage of the message id system is being able to abstract
over the actual wording of a message and just cut through to the concept
behind the logged event. Not only does this make the call sites of
RemoteIp.Debug.log/3 cleaner, it's also a short leap to see how this
could be helpful for user-facing configuration.

That is, by specifying the subset of message ids that they actually care
about, users can now customize the chattiness of RemoteIp.Debug at
compile time. Combined with the :level configuration, you could
fine-tune even your production logs to just output the couple of things
that are important to you, as demonstrated in the accompanying
integration test.