Add RemoteIp.Options
This module will eventually be used by RemoteIp to support MFA-style
options. As-is, this commit only includes the basic logical scaffold and
corresponding unit tests.

The problem is that we need to be careful about when we ultimately do
the Kernel.apply/3 that materializes an MFA into the configuration
value. For module plugs, init/1 is typically called at compile time (you
can set it to runtime, but it's not common in prod). So we don't want to
apply the MFA from there, or else it defeats the whole purpose of having
runtime semantics. E.g., environment variable lookups would just get
compiled away into whatever their values were when you ran `mix
compile`. Instead, we want to wait to apply the MFA at runtime during
each and every call/2.

However, if the user gives static (non-MFA) options to init/1, we'd be
remiss not to front-load as much work as we could at compile time. That
is, we'd want to perform redundant option parsing work (e.g.,
InetCidr.parse/1 calls) at compile time in init/1, since we could
definitely avoid doing that work at runtime during each call/2.

RemoteIp.Options therefore splits the option parsing into two phases:
"packing" and "unpacking". Packing the options will do upfront parsing
of non-MFAs but leaves MFAs untouched. Unpacking these packed options
means either directly returning the already-processed non-MFA values
(avoiding runtime work) or else applying the MFA at runtime and
processing it on the fly (work that can't be avoided, but it's okay
because that's the choice the user is making).