Use RemoteIp.Options
The one-two punch of RemoteIp.Options.pack/1 and
RemoteIp.Options.unpack/1 now lets RemoteIp support MFA-style
configuration options.

This renders the old RemoteIp.Config struct basically meaningless. It
doesn't announce "I'm the options that have already been processed"
because we might still have MFAs to contend with. Instead, all that
logic resides in the RemoteIp.Options module, which straddles the init/1
and call/2 divide. (Strictly from/2 as well, but that's kind of an
implementation detail because from/2 is necessarily already doing
everything at runtime.)

To be honest, the struct wasn't really helpful before, either. I just
introduced it in favor of ordered tuples. Keyword lists or plain maps
work just as well, though. It's not like pattern matching against a
specific "type" was ever all that useful. The only quasi-public function
RemoteIp.Config affected was call/2, effectively saying "you better make
sure to pass me options that were parsed by init/1". But that's already
understood to be the contract of module plugs, so... 🤷

In principle, we might make the code more "assertive" with some
intermediate struct returned by RemoteIp.Options.pack/1. Then we'd
pattern match against it in RemoteIp.Options.unpack/1 so that we're
encoding the rule "we can only unpack something that was previously
packed". I'm not overly concerned with doing this because the
RemoteIp.Options functions are not meant to be public anyway, and we
have good enough test coverage to be confident in its usage.

Besides which, the code just looks nicer without worrying about
RemoteIp.Config, so I'm axing it. 🪓