Make opts required in init callback
All this time, I'd been exporting an init/0 with opts defaulting to an
empty list. But the Plug behaviour only requires an init/1 function. So
the zero-arity case is just dead code, as far as I can tell.

I only noticed this because I ran `mix test --cover` and I hadn't tested
any calls of RemoteIp.init/0, so hey...here's to improved coverage! 😅