Fix dialyzer warning
Because the `@enabled` module attribute is derived at compile time, and
since the is_list and is_boolean tests are inlined by the compiler, we
wind up with a definition of enabled?/1 that's essentially equivalent to

  defp enabled?(id) do
    cond do
      false -> Enum.member?(@enabled, id)
      true -> @enabled
    end
  end

These trivial branches then generate weird-looking dialyzer warnings:

  lib/remote_ip/debug.ex:27:pattern_match The pattern can never match the type false.
  lib/remote_ip/debug.ex:28:pattern_match The pattern can never match the type true.

This commit moves the `cond` up from the runtime function body to a
compile-time check that decides which version of the function to
actually define. Thus, all the dead code is cleared away at compile
time, leaving nothing for dialyzer to complain about.