Move :package to mirror the function definition order
This is a purely cosmetic change. I'm annoyed that there isn't a more
obvious way to consistently order all the sordid configurations in the
project/0 list. But at least now the keyword pairs are in the same order
as the private function definitions that follow.

Otherwise, the ordering of project/0 goes vaguely:

1. Some defining globals (:app, :version, :elixir).
2. High-level package information (:description, :package, :deps).
3. Local development stuff (:aliases, :dialyzer, :docs).

So it reads well enough, I suppose.