Auto-format the lib directory
This is a blind application of `mix format` to the lib/ code, which is
already compliant for the most part. There are still some odds & ends to
fix by hand where the auto-formatting actually made the code look worse,
but this gets us rolling with some standards.

I'm avoiding auto-formatting the test/ directory for now because there
are significantly more issues with it, so `mix format` would make a
bunch of work I'd just have to undo.

Salient changes in this commit:

* A couple of whitespace/alignment issues were fixed. The biggest change
  is that whitespace isn't going to be used to align arrows anymore
  (e.g., in `case` statements), but I agree with that decision.

* A bunch of optional parens got re-added, mostly in typespecs.

* Apparently any calls with multiline do/end blocks are formatted such
  that the arguments preceding the `do` are NOT wrapped in parens. This
  affects the RemoteIp.Debug.log/3 call sites. Looks sort of weird to
  me, but there's no real getting around the rule, and I don't care
  enough to fight it.

* Inline comments are apparently a no-go. This won't matter too much in
  the long term, since most of them are TODO comments anyway.

* `mix format` doesn't like trailing commas. 😢 It's easier to kowtow to
  their demands rather than rock the boat. Sigh.

* Mercifully, multi-line lists/maps get their brackets put on newlines
  now. I finally don't feel compelled to use the fugly convention that
  existed circa 2016 when I first wrote this package.