Rework RemoteIp.Headers interface
While futzing around trying to figure out if there was a nicer way to do
all the extraction of known headers, I came across Keyword.take/2. We
could technically use its definition as-is, but the type spec wouldn't
be compatible with the Plug.Conn.headers/0 type. That is, headers have
the type [{binary, binary}], not [{atom, any}]. So Keyword.keyword?/1
would be false, making it weird to use Keyword.take/2 on them.

However, cribbing from its implementation, the `in` syntax is much
nicer. With the vocabulary to express the idea, I now also prefer
splitting RemoteIp.Headers.parse/2 up into two functions, take/2 and
parse/1. Single responsibility principle and all that.