Expand Elixir/OTP version combinations for CI jobs
Elixir is only officially supporting 1.7+, so we want to test 1.7
through the current stable release, 1.11.

But we get something of a combinatorial explosion because each Elixir
version can run on multiple OTP versions. For the purposes of CI, we
limit ourselves to the minimum supported Erlang/OTP as listed in
https://hexdocs.pm/elixir/master/compatibility-and-deprecations.html
However, for the latest version of Elixir, we add an additional job that
runs on the maximum supported Erlang/OTP version.

The maintainers of the erlef/setup-elixir action caution against
overly-broad version specifiers for OTP, since Erlang's versioning gets
pretty gnarly. But I'm just going to try this as a first pass, because
it's way easier to manage than setting much more granular versions.