Only use OTP > 19 for CI
Elixir 1.7 supports OTP 19-22, but the OTP 19 built on Ubuntu 18.04 has
issues with SSL, resulting in

  {:error, {:ssl, {'no such file or directory', 'ssl.app'}}}

when running `mix local.rebar --force`.

The two viable options seem to be either (1) downgrading Ubuntu or (2)
upgrading OTP.

Right now, ubuntu-latest points at 18.04 and is going to be updated to
version 20.04 soon. With the OS changing enough already, it makes me
very hesitant to use an even older version. We might parameterize the
matrix over the OS as well, but that's a lot of variables to change just
for the one job running on an Elixir version that will only be supported
for six more months.

Instead, we can simply nudge OTP up to 20 for Elixir 1.7 and call it a
day. Maybe it'll start breaking on Ubuntu 20.04, but we can cross that
bridge when we get there.