Run a separate `mix compile` step
Before, this would happen implicitly from the `mix test` task. This
change now lets us use the `--warnings-as-errors` flag to ensure that
compiling remote_ip doesn't produce any warnings.

However, the dependencies compile with a bunch of warnings already, and
I'm not going to fix any of those. So we want to make sure to run `mix
deps.compile` separately.