Cache builds between workflow runs
Because we're keying by the mix.lock file's hash, the cache should be
busted by updates to the dependencies. Otherwise, Elixir should notice
changes between the remote_ip source code and the contents of _build.

This changes the cache key for the deps directory so that the naming is
more consistent, which means that that cache will be busted on the next
CI run for this push. Not a big deal, but worth noting for posterity.