Switch to using MIX_ENV=ci during workflow
The dialyzer step is failing because the dependency was given in mix.exs
as `only: :dev`. But the workflow had been setting the MIX_ENV to "test"
globally (to avoid double-compiling for dev vs `mix test`). Thus, the
`mix dialyzer` task wouldn't exist when we tried to run it.

Rather than globally move everything to MIX_ENV=dev, I think it's better
to intentionally use a more purposeful CI environment. However, because
this affects the _build directory, I want to clear the cache as I keep
honing the workflow (lest my jobs take forever). There's no way to do
this manually for actions/cache@v2, so we have to change the keys to
force a cache bust (see https://github.com/actions/cache/issues/2).

Happily, I discovered that the ${{ runner.os }} metadata doesn't expand
to something like "ubuntu-18.04", just to "Linux". So it turns out to
not be so important for the cache at all - I'm not using any non-Linux
OS for the integration tests. By removing it from the keys, I get the
cache-busting that I happen to want right here anyway.