Add restore-keys for PLT cache
PLT generation is expensive primarily because of the large volume of
Elixir/OTP-specific modules that Dialyzer has to comb through. So if the
mix.lock file happens to change, we don't want to have to rebuild
everything from scratch. `mix dialyzer` is smart enough to only analyze
changes from updated dependencies, which should be relatively reasonable
to compute.

So if I understand correctly, by adding a restore-key that looks in the
cache under *any* mix.lock hash, we'll grab the PLTs corresponding to
the old mix.lock, generate the new PLTs (cheaper than rebuilding from
scratch), then save the results under a new cache key corresponding to
the new mix.lock (so it'll be a cache hit the next time around).

I'm less confident that this reasoning extends to different Elixir/OTP
versions, and I don't want the parallel jobs stepping on each others'
toes or anything weird like that. So I'm leaving it at just the one
"level" of restoration.