Cache integration test deps & builds during CI
This isn't perfect, as it lumps together all the integration tests'
dependencies into one cache key. But the integration tests don't take
prohibitively long as-is, so it's alright if we have to rebuild them all
from scratch. As long as our common case is cached, which it should be
because I don't expect the tests' dependencies to be changing all that
much.

I think we do want to keep this separate from the top-level deps &
_build cache, though, because (at least in principle) the integration
tests may have independent dependencies & lock files.