Use more specific cache path for PLTs
Since the cache entry was already created, I think this is going to
break the behavior when it tries to restore the old priv/ directory
under priv/plts. So now I predict it'll look like priv/plts/plts/... and
dialyzer won't have access to its old data.

However, I think a saving grace here may be that I created the cache
under the ci branch, so once this gets merged into main the cache is
going to be rebuilt anyway for the new branch - this time with the right
directory structure. Or at least that's what I hope will happen. I may
need to change the cache key later anyway. 😓 But I think it's better
than restoring the entire priv/ directory in the name of restoring PLTs,
which seems like it could cause some unexpected and dangerous behavior
down the line (i.e., if anything else starts going in priv). At the very
least, it's semantically better.