Enable workflow on pushes & pulls to main
Now that I seem to have the workflow configuration bashed out, I can go
ahead and enable CI on PRs as well as pushes. But I'll restrict it only
to the main branch, just in case I want to do some experimental
development on some other branch in the future. I get to test out if
this works one last time by pushing to the 'ci' branch. :)