Reimagine the debugging interface
The RemoteIp.Debug module was reworked in these key ways:

* It was renamed to RemoteIp.Debugger
* log/3 was renamed debug/3
* The __using__/1 callback now imports the module

The initial motivation for this came from RemoteIp.Headers and
RemoteIp.Options. The RemoteIp.Headers module is useful to keep public,
because I could see someone wanting to call parse/1 ad hoc. I also think
RemoteIp.Options would be useful for documenting how to configure
RemoteIp via the `@moduledoc`. But if it's publicly documented, that may
increase the chance someone would call its functions from outside the
RemoteIp.{call,from}/2 cycle, even if they're `@doc false` or otherwise
flagged as "don't do this".

Regardless, I think it would be unexpected to generate log statements
from calling, say, RemoteIp.Headers.take/2. This is especially true
since the messaging is kind of assuming it's being called from RemoteIp.
So I wanted to move all the debug information solely into the RemoteIp
module.

In doing so, the code got clunky with fully-qualified
RemoteIp.Debug.log/3 calls. Shit was hard to fit on single lines. Thus
the __using__/1 callback imports the macros so we can just spell it
without the module name.

But when I started using an unqualified log/3, the call sites looked
kind of confusing. The do-block based structure is inconsistent with the
Logger statements you'd probably (and rightly) associate with the word
"log". I decided a better unqualified name would be debug/3.

Noticing the naming pattern of Logger.log/3, the name Debug.debug/3 also
started looking more awkward. I renamed the module to Debugger, which
makes the `use` declaration in RemoteIp read better as well.

A benefit of this renewed interface is that the debug/3 calls are
agnostic to the backing implementation. They could be generating log
statements, telemetry events, or something else entirely. It nicely
abstracts over any number of reasonable things a debugger might do.

As a cosmetic change, the private RemoteIp.Debugger function enabled?/1
was renamed to debug?/1, along with the `@enabled` module attribute.
This makes them match the name of the app config key, so it's perhaps a
little better to read.