Print summary after integration tests
It's easy to lose your place in the output of the integration tests,
since there are a lot of messages about the dependencies being
fetched/compiled (which we don't necessarily want to silence). Rather
than having to `echo $?` every time I run `mix integrate`, I decided to
add a helpful summary.

Note that the use of the `with` statement now guarantees that we'll stop
in the middle of the mix commands if one of them fails. Not only is this
safer than before, it guarantees we only ever append one flag per app to
the summary.