Add RemoteIp.Parser behaviour
This general interface paves the way for allowing users to customize the
parsers for particular headers. The idea being that they can implement
their own module with `@behaviour RemoteIp.Parser`.

The next steps will involve actually exposing a configuration option to
the user that lets them inject whatever parser they implemented.

As a random aside, while writing the typespecs for RemoteIp.Parser I
noticed that the Plug.Conn.headers/0 type uses binary(), not String.t(),
so I thought it'd be prudent to bring the relevant typespecs inline -
hence the change in RemoteIp.Headers.