Combine :call & :from debug events into :ip
Suppose you'd configured

  config :remote_ip, debug: [:ips, :call]

Then RemoteIp.call/2 would generate logs like

  [debug] Parsed IPs from forwarding headers: ...
  [debug] Processed remote IP, found client ... to replace ...

With the same configuration, RemoteIp.from/2 would NOT generate

  [debug] Processed remote IP, found client ...

which is the intention. However, the first message would still be
generated *regardless* of whether we were using call/2 or from/2.

Since we can't truly distinguish the call/2 logs from the from/2 logs in
any of the upstream debug events, it doesn't make sense for :from and
:call to be separate.