Remove debugger's :level configuration
At most, this configuration isn't general enough. We can customize which
debug events get fired, but we can't customize their *individual* log
levels. So the :level option is less "you can configure anything!" and
more "nuclear option".

But to be honest, even adding a per-event level configuration is not
appealing. It smells a lot like over-engineering (and that's coming from
me, of all people). If I can't see the utility in per-event log levels,
it's not hard to kill this darling.

Plus, using the debug level is consistent with the name
RemoteIp.Debugger, and removing the level logic simplifies the code
significantly.