Yank RemoteIp.Block.encode/1 out of the loop
This closes the gap dramatically by eliminating the redundant
tuple-to-bitstring conversion within the loop performed by the
benchmark. Now, instead of the remote_ip check being > 1.5x slower than
inet_cidr, it's generally < 1.1x slower (and sometimes faster).

It's more of a tweak to the benchmark than to the RemoteIp.Block code,
but it underscores an important implementation consideration. That is,
we may want to encode the conn.remote_ip once before passing it along to
RemoteIp.Block.contains?/2 to avoid the redundant work, since encoding
the IP turns out to take some appreciable time.