Use integers instead of bitstrings
This makes remote_ip the fastest one in the benchmarks. 🎉

It turns out that Erlang's binaries are a boxed data type that gets
allocated on the heap, whereas integers have the potential to fit within
a word in the VM. This depends on the architecture, since data still has
to be tagged with 4 bits per word. So on a 32-bit architecture we can
use 28-bit immediate integers without spilling onto the heap, which
accommodates a healthy chunk of the IPv4 range. On 64-bit architectures
we have 60 bits to play with, which encompasses IPv4 as well as some of
the IPv6 range.

I'm also guessing that remote_ip might also be faster than cider in the
fail-fast case of mismatched protocols. cider doesn't account for this,
so not only does it incorrectly compare IPv4 to IPv6 addresses (leading
to false positives), it'll still waste cycles on the bit-math. That's my
theory-crafting, anyway.