Clean up RemoteIp.Parser.parse/1 callback
The guard clauses here weren't *really* doing anything that isn't
already expressed by the typespec from RemoteIp.Parser. True, it would
actually raise a FunctionClauseError when given a non-string argument,
but (a) it's not the most semantic error and (b) it seems like an
unnecessarily slippery slope to get into the game of runtime type
checking.

I also changed the name of the parameter on the callback, since it
matches the names used by the actual implementations. If it's good
enough for them, it's good enough for the typespec.