Update integration tests' mix.lock files
These are the result of simply running `mix integrate` locally with the
new mix.exs/mix.lock file that remote_ip has.

If we didn't commit this, it wouldn't break CI or anything, but it would
keep generating messages like

    --> mix deps.get
    Resolving Hex dependencies...
    Resolution completed in 0.027s
    Unchanged:
      combine 0.10.0
    Upgraded:
      mime 1.5.0 => 2.0.3 (major)
      plug 1.11.0 => 1.14.0
      plug_crypto 1.2.0 => 1.2.3
      telemetry 0.4.2 => 1.1.0 (major)
    * Updating plug (Hex package)
    * Updating mime (Hex package)
    * Updating plug_crypto (Hex package)
    * Updating telemetry (Hex package)

CI also caches the integration test dependencies based on the hash of
their mix.lock files, so it's worth keeping them up to date for that
reason, I suppose.