Try out OTP 23 on Elixir 1.12 & 1.13 in CI
Tests are passing, but CI is failing due to the old excoveralls issue
seen before in https://github.com/parroty/excoveralls/issues/254

However, the latest workarounds noted on that issue are for earlier
versions of excoveralls that still used hackney instead of httpc (cf.
https://github.com/parroty/excoveralls/pull/311). So it's not clear that
there are any good/known workarounds, as in the old issue. In
particular, we should be running the latest castore version, and I don't
think installing/updating certifi is going to do anything.

The tests seem to pass on the other Elixir versions in the matrix, but
they're all running at least OTP 23. So I'm wondering if OTP 23 would
unstick Elixir 1.12 & 1.13. Not ideal, but worth a shot.